% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mc.R
\name{estimate_mc_prev}
\alias{estimate_mc_prev}
\title{Run a model fit for one country}
\usage{
estimate_mc_prev(
  isocode,
  tiffname,
  csvname,
  year_first = 1970,
  year_final = 2030
)
}
\arguments{
\item{isocode}{ISO-3166 3-letter alphabetic code for the country to be fitted (e.g., "UGA" for Uganda)}

\item{tiffname}{Figure name for output plot showing goodness of fit. This should end in .tiff}

\item{csvname}{File name for an output CSV of male circumcision prevalence point estimates}

\item{first_year}{First year of estimates}

\item{final_year}{Final year of estimates}
}
\description{
Estimate male circumcision prevalence and produce a diagnostic plot showing
goodness-of-fit to available survey data, as well as point estimates of
male circumcision prevalence trends by five-year age group (15-19, 20-24, ..., 45-49)
}
