% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mc.R
\name{fit_mc_model}
\alias{fit_mc_model}
\title{Fit the male circumcision uptake model for one country}
\usage{
fit_mc_model(pop_data, svy_data)
}
\arguments{
\item{pop_data}{A long data frame of male population sizes by year and single age for the selected country}

\item{svy_data}{A data frame of male circumcision prevalence estimates from household surveys for the selected country}
}
\value{
An model fit object (See details section)
}
\description{
Fit the male circumcision uptake model for one country
}
\section{Details}{


The model fit object is a list of the following:
\enumerate{
\item{resample} A matrix of posterior parameter values, one row per sample, one column per parameter
\item{posterior} Posterior density at each sample. This is calculated by Stan, and accounts for
non-negativity constraints on parameters, which will cause the values to differ from those calculated
using \code{prior()} + \code{likelihood()}.
\item{raw} Raw output from \code{stan()}. Useful for checking convergence diagnostics.
}
}

