% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mc.R
\name{fit_model}
\alias{fit_model}
\title{Fit the male circumcision uptake model for one country}
\usage{
fit_model(pop_data, svy_data)
}
\arguments{
\item{pop_data}{A long data frame of male population sizes by year and single age for the selected country}

\item{svy_data}{A data frame of male circumcision prevalence estimates from household surveys for the selected country}
}
\value{
An IMIS model fit object (See details section)
}
\description{
Fit the male circumcision uptake model for one country
}
\section{Details}{


The IMIS model fit object is a list of the following:
\enumerate{
\item{resample} A matrix of posterior parameter values, one row per sample, one column per parameter
\item{center} IMIS constructs a gaussian mixture model approximation to the posterior. "center" is a matrix of
modes of the gaussian components of that mixture model
\item{stat} Convergence monitoring statistics from IMIS iterations. These are also displayed in the console
when fit_model is running
\item{prior} log prior density at each resample
\item{lhood} log likelihood at each resample
}
}

