% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-defs.R
\name{plot_fitted_mc_rates}
\alias{plot_fitted_mc_rates}
\title{Plot fitted male circumcision uptake rates}
\usage{
plot_fitted_mc_rates(tiffname, imis_fit, pop_data)
}
\arguments{
\item{tiffname}{Output figure filename (tiff format)}

\item{imis_fit}{IMIS model fit object}

\item{pop_data}{A long data frame of male population sizes by year and single
age for the selected country}
}
\description{
Plot fitted male circumcision uptake rates for a single country as a TIFF file
}
