% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-defs.R
\name{write_mc_prev}
\alias{write_mc_prev}
\title{Write male circumcision prevalence trends to a CSV file}
\usage{
write_mc_prev(csvname, imis_fit, pop_data)
}
\arguments{
\item{csvname}{File name for an output CSV of male circumcision prevalence point estimates}

\item{imis_fit}{IMIS model fit object}

\item{pop_data}{A long data frame of male population sizes by year and single
age for the selected country}
}
\description{
Write male circumcision prevalence trends to a CSV file
}
